Restructured MainContent to /pages. Created RightPreview components. Restructured MainContent to /pages. Created RightPreview components.
