+++ components/layouts/DashboardLayout.js
@@ -2,12 +2,12 @@ import React from "react";
 import RightPreview from "../dashboard/RightPreview";
 import Sidebar from "../dashboard/Sidebar";
 
-const DashboardLayout = ({ children }) => {
+const DashboardLayout = ({ children, userdata }) => {
   return (
     <div className="main-wrapper flex flex-row h-screen w-screen overflow-y-auto">
-      <Sidebar />
+      <Sidebar userdata={userdata} />
       <div className="w-full">{children}</div>
-      <RightPreview />
+      <RightPreview userdata={userdata} />
     </div>
   );
 };
